update delete post function

+++ backend/BakeryRecipe/src/main/java/com/example/bakeryrecipe/api/PostAPI.java
@@ -3,9 +3,11 @@ package com.example.bakeryrecipe.api;
 import com.example.bakeryrecipe.dto.PostDTO;
 import com.example.bakeryrecipe.service.PostService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PostAuthorize;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -32,8 +34,13 @@ public class PostAPI {
     @PostAuthorize("#returnObject.member.id.equals(authentication.principal.id)")
     @DeleteMapping("/{id}")
     public PostDTO deletePost(@PathVariable("id") Long id){
+        PostDTO postDTO = postService.delete(id);
+        if(postDTO != null){
+            return new ResponseStatusException(HttpStatus.NOT_FOUND,"not found post");
+        }
         return postService.delete(id);
     }
+
 
     @GetMapping
     public List<PostDTO> listPost(){
